# Speed up `countDigits`