--- a/tests/performance/countDigits.xml
+++ b/tests/performance/countDigits.xml
@@ -1,0 +1,6 @@
+<test>
+    <query>SELECT count() FROM test.hits WHERE NOT ignore(countDigits(RegionID))</query>
+    <query>SELECT count() FROM test.hits WHERE NOT ignore(countDigits(AdvEngineID))</query>
+    <query>SELECT count() FROM test.hits WHERE NOT ignore(countDigits(ClientIP))</query>
+    <query>SELECT count() FROM test.hits WHERE NOT ignore(countDigits(WatchID))</query>
+</test>