… tests/queries/0_stateless/01458_count_digits.reference
… tests/queries/0_stateless/02560_count_digits.reference