### Changelog category (leave one):
- Backward Incompatible Change


### Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):
Slightly improve performance of `countDigits` on realistic datasets. This closed #44518. In previous versions, `countDigits(0)` returned `0`; now it returns `1`, which is more correct, and follows the existing documentation.


Continuation of #44518.