@alexey-milovidov 2023-02-09
Before:
```
milovidov@milovidov-desktop:~$ for column in {RegionID,ClientIP,WatchID,AdvEngineID}; do clickhouse-benchmark -i100 --query "SELECT sum(countDigits($column)) FROM hits"; done
Loaded 1 queries.

Queries executed: 100.

localhost:9000, queries 100, QPS: 36.253, RPS: 3625209320.385, MiB/s: 13829.076, result RPS: 36.253, result MiB/s: 0.000.

0.000%          0.024 sec.
10.000%         0.025 sec.
20.000%         0.025 sec.
30.000%         0.026 sec.
40.000%         0.026 sec.
50.000%         0.026 sec.
60.000%         0.027 sec.
70.000%         0.027 sec.
80.000%         0.028 sec.
90.000%         0.029 sec.
95.000%         0.033 sec.
99.000%         0.054 sec.
99.900%         0.090 sec.
99.990%         0.090 sec.

Queries executed: 100.

localhost:9000, queries 100, QPS: 34.905, RPS: 3490418577.064, MiB/s: 13314.890, result RPS: 34.905, result MiB/s: 0.000.

0.000%          0.026 sec.
10.000%         0.026 sec.
20.000%         0.027 sec.
30.000%         0.027 sec.
40.000%         0.027 sec.
50.000%         0.028 sec.
60.000%         0.028 sec.
70.000%         0.029 sec.
80.000%         0.030 sec.
90.000%         0.031 sec.
95.000%         0.032 sec.
99.000%         0.036 sec.
99.900%         0.066 sec.
99.990%         0.066 sec.

Queries executed: 100.

localhost:9000, queries 100, QPS: 20.091, RPS: 2009053467.947, MiB/s: 15327.862, result RPS: 20.091, result MiB/s: 0.000.

0.000%          0.047 sec.
10.000%         0.048 sec.
20.000%         0.048 sec.
30.000%         0.049 sec.
40.000%         0.049 sec.
50.000%         0.049 sec.
60.000%         0.049 sec.
70.000%         0.050 sec.
80.000%         0.051 sec.
90.000%         0.052 sec.
95.000%         0.053 sec.
99.000%         0.056 sec.
99.900%         0.073 sec.
99.990%         0.073 sec.


Queries executed: 100.

localhost:9000, queries 100, QPS: 58.320, RPS: 5831856159.607, MiB/s: 11123.383, result RPS: 58.320, result MiB/s: 0.000.

0.000%          0.014 sec.
10.000%         0.015 sec.
20.000%         0.015 sec.
30.000%         0.016 sec.
40.000%         0.017 sec.
50.000%         0.017 sec.
60.000%         0.017 sec.
70.000%         0.017 sec.
80.000%         0.018 sec.
90.000%         0.018 sec.
95.000%         0.019 sec.
99.000%         0.036 sec.
99.900%         0.044 sec.
99.990%         0.044 sec.
```

After:

```
milovidov@milovidov-desktop:~$ for column in {RegionID,ClientIP,WatchID,AdvEngineID}; do clickhouse-benchmark -i100 --query "SELECT sum(countDigits($column)) FROM hits"; done
Loaded 1 queries.

Queries executed: 100.

localhost:9000, queries 100, QPS: 42.361, RPS: 4235954485.892, MiB/s: 16158.884, result RPS: 42.361, result MiB/s: 0.000.

0.000%          0.019 sec.
10.000%         0.021 sec.
20.000%         0.021 sec.
30.000%         0.022 sec.
40.000%         0.022 sec.
50.000%         0.022 sec.
60.000%         0.023 sec.
70.000%         0.023 sec.
80.000%         0.024 sec.
90.000%         0.027 sec.
95.000%         0.030 sec.
99.000%         0.051 sec.
99.900%         0.053 sec.
99.990%         0.053 sec.

Queries executed: 100.

localhost:9000, queries 100, QPS: 42.068, RPS: 4206646165.442, MiB/s: 16047.082, result RPS: 42.068, result MiB/s: 0.000.

0.000%          0.020 sec.
10.000%         0.021 sec.
20.000%         0.021 sec.
30.000%         0.022 sec.
40.000%         0.022 sec.
50.000%         0.023 sec.
60.000%         0.023 sec.
70.000%         0.024 sec.
80.000%         0.025 sec.
90.000%         0.027 sec.
95.000%         0.029 sec.
99.000%         0.044 sec.
99.900%         0.048 sec.
99.990%         0.048 sec.

Queries executed: 100.

localhost:9000, queries 100, QPS: 19.724, RPS: 1972324306.630, MiB/s: 15047.640, result RPS: 19.724, result MiB/s: 0.000.

0.000%          0.048 sec.
10.000%         0.049 sec.
20.000%         0.049 sec.
30.000%         0.050 sec.
40.000%         0.050 sec.
50.000%         0.050 sec.
60.000%         0.051 sec.
70.000%         0.051 sec.
80.000%         0.052 sec.
90.000%         0.053 sec.
95.000%         0.054 sec.
99.000%         0.057 sec.
99.900%         0.065 sec.
99.990%         0.065 sec.

Queries executed: 100.

localhost:9000, queries 100, QPS: 65.619, RPS: 6561708923.538, MiB/s: 12515.467, result RPS: 65.619, result MiB/s: 0.001.

0.000%          0.012 sec.
10.000%         0.012 sec.
20.000%         0.014 sec.
30.000%         0.015 sec.
40.000%         0.015 sec.
50.000%         0.015 sec.
60.000%         0.015 sec.
70.000%         0.015 sec.
80.000%         0.016 sec.
90.000%         0.017 sec.
95.000%         0.020 sec.
99.000%         0.023 sec.
99.900%         0.036 sec.
99.990%         0.036 sec.
```